refactor(headroom): move managed MCP install into internal/headroom
